# docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Untripical API",
        "description": "API aplication Untripical",
        "contact": {
            "name": "Mateusz",
            "email": "mateusz@untripical.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Untripical server"
        }
    ],
    "paths": {
...
        "/announcement/update/{name}": {
            "put": {
                "tags": [
                    "announcement-controller"
                ],
                "operationId": "updateDescription",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnnouncementUpdateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnnouncementResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/announcement/add": {
            "post": {
                "tags": [
                    "announcement-controller"
                ],
                "operationId": "addAnnouncement",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnnouncementRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnnouncementResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/announcement/type/{announcementType}": {
            "get": {
                "tags": [
                    "announcement-controller"
                ],
                "operationId": "getAnnouncementsByAnnouncementType",
                "parameters": [
                    {
                        "name": "announcementType",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "GROUP",
                                "INDIVIDUAL"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AnnouncementResponseDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/announcement/placeId/{placeId}": {
            "get": {
                "tags": [
                    "announcement-controller"
                ],
                "operationId": "getAnnouncementsByPlaceId",
                "parameters": [
                    {
                        "name": "placeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AnnouncementResponseDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/announcement/location/{locationInfo}": {
            "get": {
                "tags": [
                    "announcement-controller"
                ],
                "operationId": "getAnnouncementsByLocationInfo",
                "parameters": [
                    {
                        "name": "locationInfo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AnnouncementResponseDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/announcement/id/{id}": {
            "get": {
                "tags": [
                    "announcement-controller"
                ],
                "operationId": "getAnnouncementById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnnouncementResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/announcement/date": {
            "get": {
                "tags": [
                    "announcement-controller"
                ],
                "operationId": "getAnnouncementsByDate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnnouncementDateRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AnnouncementResponseDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/announcement/delete/{name}": {
            "delete": {
                "tags": [
                    "announcement-controller"
                ],
                "operationId": "deleteAnnouncement",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnnouncementResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "AnnouncementResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "nameOfJourney": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "announcementType": {
                        "type": "string",
                        "enum": [
                            "GROUP",
                            "INDIVIDUAL"
                        ]
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    },
                    "locationInfo": {
                        "type": "string"
                    },
                    "maxParticipants": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "placeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "guideAnnouncementTables": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GuideAnnouncementTableResponseDTO"
                        }
                    }
                }
            },
            "GuideAnnouncementTableResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "announcementId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "guideDetailsId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "AnnouncementUpdateDTO": {
                "type": "object",
                "properties": {
                    "nameOfJourney": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "announcementType": {
                        "type": "string",
                        "enum": [
                            "GROUP",
                            "INDIVIDUAL"
                        ]
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "price": {
                        "type": "number",
                        "format": "double",
                        "minimum": 0
                    },
                    "locationInfo": {
                        "type": "string"
                    },
                    "maxParticipants": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "placeId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "AnnouncementRequestDTO": {
                "type": "object",
                "properties": {
                    "nameOfJourney": {
                        "type": "string",
                        "minLength": 1
                    },
                    "description": {
                        "type": "string",
                        "minLength": 1
                    },
                    "announcementType": {
                        "type": "string",
                        "enum": [
                            "GROUP",
                            "INDIVIDUAL"
                        ]
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "price": {
                        "type": "number",
                        "format": "double",
                        "minimum": 0
                    },
                    "locationInfo": {
                        "type": "string",
                        "minLength": 1
                    },
                    "maxParticipants": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "placeId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "announcementType",
                    "date",
                    "maxParticipants",
                    "placeId",
                    "price"
                ]
            },
            "AnnouncementDateRequestDTO": {
                "type": "object",
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "required": [
                    "date"
                ]
            }
        }
    }
}